requirements.txt
opencv-python>=4.8.0.74
mediapipe>=0.10.14
numpy>=1.26.0
matplotlib>=3.8.0
scipy>=1.11.0
pandas>=2.0.0
tqdm>=4.66.0
streamlit>=1.36.0
yt-dlp>=2024.8.6
jinja2>=3.1.4
pyyaml>=6.0.1
